requirements.txt
pymorphy2~=0.8
pytz~=2020.1
requests~=2.23.0
psycopg2-binary~=2.8.5